Group endpoints by category
This splits `ListenBrainz.Endpoints.cs` into multiple files, grouped by
the categories used in the official ListenBrainz API docs.

This means that currently only the `Core` and `Statistics` parts contain
actual code, but it makes it easier to track missing endpoints as they
get added.
