+++ MetaBrainz.ListenBrainz/ListenBrainz.EndPoints.Art.cs
@@ -1,0 +1,5 @@
+namespace MetaBrainz.ListenBrainz;
+
+public sealed partial class ListenBrainz {
+
+}
